Add tour model
Will be used to store tours created by users. Right now
we're going to rely on the socrata api to store all of the
data regarding landmarks, so the tours will only stor a json list
of landmarks in the order the were saved.

This may change in the future. We might store this data in a more structured format, but this will keep our options open for now.